[storage] UBA ChunkedMmappedFile/CompressedMmappedFile: fix Wshorted warnings

+++ src/storage/UByteArrayAdapterPrivates/UByteArrayAdapterPrivateChunkedMmappedFile.cpp
@@ -161,7 +161,7 @@ UByteArrayAdapter::OffsetType UByteArrayAdapterPrivateChunkedMmappedFile::getOff
 
 uint64_t UByteArrayAdapterPrivateChunkedMmappedFile::getVlPackedUint64(UByteArrayAdapter::OffsetType pos, int * length) const {
 	SizeType bufLen = (*length > 9 ? 9 : *length);
-	*length = bufLen;
+	*length = (int) bufLen;
 	uint8_t buf[bufLen];
 #ifdef SSERIALIZE_WITH_THREADS
 	m_fileLock.lock();
@@ -176,7 +176,7 @@ uint64_t UByteArrayAdapterPrivateChunkedMmappedFile::getVlPackedUint64(UByteArra
 
 int64_t UByteArrayAdapterPrivateChunkedMmappedFile::getVlPackedInt64(UByteArrayAdapter::OffsetType pos, int * length) const {
 	SizeType bufLen = (*length > 9 ? 9 : *length);
-	*length = bufLen;
+	*length = (int) bufLen;
 	uint8_t buf[bufLen];
 #ifdef SSERIALIZE_WITH_THREADS
 	m_fileLock.lock();
@@ -190,7 +190,7 @@ int64_t UByteArrayAdapterPrivateChunkedMmappedFile::getVlPackedInt64(UByteArrayA
 
 uint32_t UByteArrayAdapterPrivateChunkedMmappedFile::getVlPackedUint32(UByteArrayAdapter::OffsetType pos, int * length) const {
 	SizeType bufLen = (*length > 5 ? 5 : *length);
-	*length = bufLen;
+	*length = (int) bufLen;
 	uint8_t buf[bufLen];
 #ifdef SSERIALIZE_WITH_THREADS
 	m_fileLock.lock();
